refactor: extract useNetworkState hook from NetworkService
Move the React hook into its own file under src/hooks/useNetworkState.ts.
Remove React import from the service file. Remove broken
ErrorHandler.updateNetworkState call (method was removed in #9).
Update all imports to use the new hook location.

Closes #7
