+++ NaturalWineDetector/src/hooks/index.ts
@@ -1,6 +1,7 @@
 // Export all custom hooks from this file
 export * from './useLocation';
 export * from './useCamera';
+export * from './useNetworkState';
 export * from './useWineRepository';
 export * from './usePermissions';
 export * from './usePerformanceMonitor';
